requirements.txt
flask
google-generativeai
requests
flask-cors